fix type of date

+++ src/utils/autoRun.js
@@ -17,7 +17,7 @@ exports.mqtt = async (topic, message) => {
       {
         $push: {
           data: {
-            value: parseInt(message),
+            value: parseFloat(message),
             createAt: (new Date()).toString()
           }
         }
